Fix renaming bugs
