+++ src/plugins/index.browser.ts
@@ -1,2 +1,3 @@
 // # index.browser.ts
 export { default as PluginIndex } from './plugin-index.browser.js';
+export type { default as DefaultPluginIndex } from './plugin-index.js';
