+++ src/plugins/index.ts
@@ -7,3 +7,4 @@ export { default as createSubmenuPatch, SubmenuPatcher } from './create-submenu-
 export { default as createSubmenuButton } from './create-submenu-button.js';
 export { default as unpackSubmenu } from './unpack-submenu.js';
 export { default as createLoadComparator } from './create-load-comparator.js';
+export type { default as BasePluginIndex } from './plugin-index.js';
